+++ includes/class.GFEwayFormData.php
@@ -9,8 +9,6 @@ if (!defined('ABSPATH')) {
 */
 class GFEwayFormData {
 
-	public $amount					= 0;
-	public $shipping				= 0;
 	public $total					= 0;
 	public $ccName					= '';
 	public $ccNumber				= '';
@@ -57,7 +55,8 @@ class GFEwayFormData {
 		foreach ($form['fields'] as &$field) {
 			$id = $field->id;
 
-			switch(GFFormsModel::get_input_type($field)){
+			switch (GFFormsModel::get_input_type($field)) {
+
 				case 'name':
 					// only pick up the first name field (assume later ones are additional info)
 					if (empty($this->firstName) && empty($this->lastName)) {
@@ -112,11 +111,6 @@ class GFEwayFormData {
 					break;
 
 				case 'total':
-					$this->total					= GFCommon::to_number(rgpost("input_{$id}"));
-					// handle condition where Total field comes back blank (e.g. when T2T Toolkit is messing up Gravity Forms) or invalid
-					if ($this->total === false) {
-						$this->total = 0;
-					}
 					$this->hasPurchaseFieldsFlag	= true;
 					break;
 
@@ -128,125 +122,16 @@ class GFEwayFormData {
 					break;
 
 				default:
-					// check for shipping field
-					if ($field->type === 'shipping') {
-						$this->shipping += self::getShipping($form, $field);
-						$this->hasPurchaseFieldsFlag = true;
-					}
-					// check for product field
-					elseif (self::isProductField($field)) {
-						$this->amount += self::getProductPrice($form, $field);
+					if ($field->type === 'shipping' || $field->type === 'product') {
 						$this->hasPurchaseFieldsFlag = true;
 					}
 					break;
+
 			}
 		}
 
-		// if form didn't pass the total, use sum of the product and shipping fields
-		if ($this->total === 0) {
-			$this->total = $this->amount + $this->shipping;
-		}
-	}
-
-	/**
-	* extract the price from a product field, and multiply it by the quantity
-	* @return float
-	*/
-	private static function getProductPrice($form, $field) {
-		$price = $qty = 0;
-		$isProduct = false;
-		$id = $field->id;
-
-		if (!GFFormsModel::is_field_hidden($form, $field, array())) {
-			$lead_value = rgpost("input_{$id}");
-
-			// look for a quantity field for product
-			$qty_fields = GFCommon::get_product_fields_by_type($form, array('quantity'), $id);
-			if (empty($qty_fields)) {
-				$qty_field = false;
-				$qty = 1;
-			}
-			else {
-				$qty_field = $qty_fields[0];
-				$qty = (float) rgpost("input_{$qty_field['id']}");
-			}
-
-			switch ($field->inputType) {
-				case 'singleproduct':
-				case 'hiddenproduct':
-					$price = GFCommon::to_number(rgpost("input_{$id}_2"));
-					if (!$qty_field) {
-						// no quantity field, pick it up from input
-						$qty = (float) GFCommon::to_number(rgpost("input_{$id}_3"));
-					}
-					$isProduct = true;
-					break;
-
-				case 'donation':
-				case 'price':
-					$price = GFCommon::to_number($lead_value);
-					$isProduct = true;
-					break;
-
-				default:
-					// handle drop-down lists
-					if (!empty($lead_value)) {
-						list($name, $price) = rgexplode('|', $lead_value, 2);
-						$isProduct = true;
-					}
-					break;
-			}
-
-			// pick up extra costs from any options
-			if ($isProduct) {
-				$options = GFCommon::get_product_fields_by_type($form, array('option'), $id);
-				foreach($options as $option){
-					if (!GFFormsModel::is_field_hidden($form, $option, array())) {
-						$option_value = rgpost("input_{$option['id']}");
-
-						if (is_array(rgar($option, 'inputs'))) {
-							foreach($option['inputs'] as $input){
-								$input_value = rgpost('input_' . str_replace('.', '_', $input['id']));
-								$option_info = GFCommon::get_option_info($input_value, $option, true);
-								if(!empty($option_info))
-									$price += GFCommon::to_number(rgar($option_info, 'price'));
-							}
-						}
-						elseif (!empty($option_value)){
-							$option_info = GFCommon::get_option_info($option_value, $option, true);
-							$price += GFCommon::to_number(rgar($option_info, 'price'));
-						}
-					}
-				}
-
-				$price *= $qty;
-			}
-
-		}
-
-		return $price;
-	}
-
-	/**
-	* extract the shipping amount from a shipping field
-	* @return float
-	*/
-	private static function getShipping($form, $field) {
-		$shipping = 0;
-		$id = $field->id;
-
-		if (!GFFormsModel::is_field_hidden($form, $field, array())) {
-			$value = rgpost("input_{$id}");
-
-			if (!empty($value) && $field->inputType != 'singleshipping') {
-				// drop-down list / radio buttons
-				list($name, $value) = rgexplode('|', $value, 2);
-			}
-
-			$shipping = GFCommon::to_number($value);
-		}
-
-		return $shipping;
+		$entry = GFFormsModel::get_current_lead();
+		$this->total = GFCommon::get_order_total($form, $entry);
 	}
 
 	/**
@@ -288,17 +173,6 @@ class GFEwayFormData {
 	*/
 	public function hasRecurringPayments() {
 		return !!$this->recurring;
-	}
-
-	/**
-	* test if field is a product field
-	* @param GFField $field
-	* @return bool
-	*/
-	public static function isProductField($field) {
-		static $productTypes = array('option', 'donation', 'product', 'calculation');
-
-		return in_array($field->type, $productTypes);
 	}
 
 }